Refactored so that CameraSystem and AndroidCameraSystem could be reduced to just a CameraSystem that doesn't know about Android stuff.
