+++ packages/camera/android/src/main/java/dev/flutter/plugins/camera/CameraFactory.java
@@ -1,0 +1,14 @@
+package dev.flutter.plugins.camera;
+
+import android.hardware.camera2.CameraAccessException;
+
+import androidx.annotation.NonNull;
+
+public interface CameraFactory {
+  @NonNull
+  Camera createCamera(
+      @NonNull String cameraName,
+      @NonNull String resolutionPreset,
+      boolean enableAudio
+  ) throws CameraAccessException;
+}
