+++ packages/camera/android/src/main/java/dev/flutter/plugins/camera/CameraHardware.java
@@ -1,0 +1,12 @@
+package dev.flutter.plugins.camera;
+
+import android.hardware.camera2.CameraAccessException;
+
+import androidx.annotation.NonNull;
+
+import java.util.List;
+
+public interface CameraHardware {
+  @NonNull
+  List<CameraDetails> getAvailableCameras() throws CameraAccessException;
+}
